Extract the adb path from the project configuration
