+++ core/src/main/scala/com/karumi/shot/Adb.scala
@@ -1,0 +1,9 @@
+package com.karumi.shot
+
+object Adb {
+  var adbBinaryPath: String = ""
+}
+class Adb {
+
+
+}
