+++ core/src/main/scala/com/karumi/shot/Shot.scala
@@ -1,3 +1,9 @@
 package com.karumi.shot
 
-class Shot {}
+class Shot(val adb: Adb) {
+
+  def configureAdbPath(adbPath: String) {
+    Adb.adbBinaryPath = adbPath
+  }
+
+}
