--- core/src/main/scala/com/karumi/shot/app.scala
@@ -1,3 +1,0 @@
-package com.karumi.shot
-
-object app {}
