+++ shot/src/main/groovy/com/karumi/shot/AdbPathExtractor.groovy
@@ -1,0 +1,10 @@
+package com.karumi.shot
+
+import org.gradle.api.Project
+
+class AdbPathExtractor {
+
+    static String extracPath(Project project) {
+        project.android.getAdbExe().toString()
+    }
+}
